Checkpoint System
The checkpoint system is a feature in the platform game I've been working on which saves the players progress at key locations, as well as providing a location for the player to respawn at. It uses ingame world coordinates which are usually tied to an ingame object that gives the player a visual for where the checkpoint exists.